requirements.txt
streamlit 
python-dotenv
google-generativeai
langchain
PYPDF2
chromadb